Add "readonly mode" to feature, feature definition and remove delete menu in menustips menu
git-svn-id: https://spm.svn.codeplex.com/svn@54280 81fbe566-5dc4-48c1-bdea-7421811ca204

+++ SharePoint Manager 2010/SharePoint Manager 2007/Components/Menu/MenuStripStandard.cs
@@ -23,6 +23,10 @@ namespace Keutmann.SharePointManager.Components
         public ToolStripItem DeleteItem;
         public ToolStripItem RefreshItem;
 
+        public bool InReadOnlyMode
+        {
+            get { return Properties.Settings.Default.ReadOnly; }
+        }
 
         public MenuStripStandard()
         {
@@ -53,9 +57,15 @@ namespace Keutmann.SharePointManager.Components
             //    RefreshItem
             //   });
 
-            this.Items.AddRange(new ToolStripItem[] {
+            if (!InReadOnlyMode)
+            {
+                this.Items.AddRange(new ToolStripItem[] {
                 DeleteItem,
-                SPMMenu.Items.CreateSeparator(),
+                SPMMenu.Items.CreateSeparator()
+               });
+                
+            }
+            this.Items.AddRange(new ToolStripItem[] {
                 RefreshItem
                });
